requirements.txt
asgiref==3.8.1
comtypes==1.4.7
distlib==0.3.9
Django==5.1.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
emoji==2.12.1
filelock==3.16.1
mysqlclient==2.2.7
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyjokes==0.8.3
PyJWT==2.10.1
pypiwin32==223
pyttsx3==2.97
pywin32==306
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.0
